clean up and wip

+++ geowave-analytics/src/main/java/mil/nga/giat/geowave/analytics/sample/BahmanEtAlSampleProbabilityFn.java
@@ -17,7 +17,7 @@ public class BahmanEtAlSampleProbabilityFn implements
 			final double weight,
 			final double normalizingConstant,
 			final int sampleSize ) {
-		return (((double) sampleSize) * weight) / normalizingConstant;
+		return ((sampleSize) * weight) / normalizingConstant;
 	}
 
 	@Override
